Create Cards_12_Word_Seed.py
Uses 52 deck cards and 4 jokers to create random entropy and a 12 word seed. have a play around but do NOT use it for any real seed. Just learn and do a print version for a real seed 100% offline always. 

Run in a Kivy from the command prompt. Put the 2048 wordlist.txt in the same folder.