SeedMaker Cards 12 Word Seed
Uses a deck of common 52 deck playing cards and adds 4 jokers to make it unbiased and creates BIP39 Mnemonic Seed Phrases. 


WARNING do not use this script for any online wallet seed generation or any seed generation at all for that matter. Unless you know exactly what you are doing and axcept the risks. You should alway do a seed 100% offline and i suggest use the print version so it is randomly worked out using dice or cards or flipping a coin and pen and paper to work the seed out. I have provided pdf versions for this very way.